renamed and reorganize files to follow gem file structure
