+++ lib/http_codes.rb
@@ -1,2 +1,2 @@
-require 'http_codes/httpcodes.rb'
+require 'http_codes/http_codes_fetcher.rb'
 require 'http_codes/version.rb'
