--- spec/spec/fixtures/vcr_cassettes/http_csv.yml
@@ -1,82 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: http://www.iana.org/assignments/http-status-codes/http-status-codes-1.csv
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Last-Modified:
-      - Thu, 05 Nov 2015 01:13:26 GMT
-      Vary:
-      - Accept-Encoding
-      Content-Type:
-      - text/csv; charset=UTF-8; header=present
-      Cache-Control:
-      - public, max-age=3600
-      Content-Length:
-      - '1031'
-      Accept-Ranges:
-      - bytes
-      Date:
-      - Thu, 05 Nov 2015 15:23:25 GMT
-      Connection:
-      - keep-alive
-      Server:
-      - Apache
-      X-Cache-Host:
-      - cache1.lax.icann.org
-      X-Cache-Hits:
-      - '1'
-    body:
-      encoding: ASCII-8BIT
-      string: "Value,Description,Reference\r\n100,Continue,\"[RFC7231, Section 6.2.1]\"\r\n101,Switching
-        Protocols,\"[RFC7231, Section 6.2.2]\"\r\n102,Processing,[RFC2518]\r\n103-199,Unassigned,\r\n200,OK,\"[RFC7231,
-        Section 6.3.1]\"\r\n201,Created,\"[RFC7231, Section 6.3.2]\"\r\n202,Accepted,\"[RFC7231,
-        Section 6.3.3]\"\r\n203,Non-Authoritative Information,\"[RFC7231, Section
-        6.3.4]\"\r\n204,No Content,\"[RFC7231, Section 6.3.5]\"\r\n205,Reset Content,\"[RFC7231,
-        Section 6.3.6]\"\r\n206,Partial Content,\"[RFC7233, Section 4.1]\"\r\n207,Multi-Status,[RFC4918]\r\n208,Already
-        Reported,[RFC5842]\r\n209-225,Unassigned,\r\n226,IM Used,[RFC3229]\r\n227-299,Unassigned,\r\n300,Multiple
-        Choices,\"[RFC7231, Section 6.4.1]\"\r\n301,Moved Permanently,\"[RFC7231,
-        Section 6.4.2]\"\r\n302,Found,\"[RFC7231, Section 6.4.3]\"\r\n303,See Other,\"[RFC7231,
-        Section 6.4.4]\"\r\n304,Not Modified,\"[RFC7232, Section 4.1]\"\r\n305,Use
-        Proxy,\"[RFC7231, Section 6.4.5]\"\r\n306,(Unused),\"[RFC7231, Section 6.4.6]\"\r\n307,Temporary
-        Redirect,\"[RFC7231, Section 6.4.7]\"\r\n308,Permanent Redirect,[RFC7538]\r\n309-399,Unassigned,\r\n400,Bad
-        Request,\"[RFC7231, Section 6.5.1]\"\r\n401,Unauthorized,\"[RFC7235, Section
-        3.1]\"\r\n402,Payment Required,\"[RFC7231, Section 6.5.2]\"\r\n403,Forbidden,\"[RFC7231,
-        Section 6.5.3]\"\r\n404,Not Found,\"[RFC7231, Section 6.5.4]\"\r\n405,Method
-        Not Allowed,\"[RFC7231, Section 6.5.5]\"\r\n406,Not Acceptable,\"[RFC7231,
-        Section 6.5.6]\"\r\n407,Proxy Authentication Required,\"[RFC7235, Section
-        3.2]\"\r\n408,Request Timeout,\"[RFC7231, Section 6.5.7]\"\r\n409,Conflict,\"[RFC7231,
-        Section 6.5.8]\"\r\n410,Gone,\"[RFC7231, Section 6.5.9]\"\r\n411,Length Required,\"[RFC7231,
-        Section 6.5.10]\"\r\n412,Precondition Failed,\"[RFC7232, Section 4.2]\"\r\n413,Payload
-        Too Large,\"[RFC7231, Section 6.5.11]\"\r\n414,URI Too Long,\"[RFC7231, Section
-        6.5.12]\"\r\n415,Unsupported Media Type,\"[RFC7231, Section 6.5.13][RFC7694,
-        Section 3]\"\r\n416,Range Not Satisfiable,\"[RFC7233, Section 4.4]\"\r\n417,Expectation
-        Failed,\"[RFC7231, Section 6.5.14]\"\r\n418-420,Unassigned,\r\n421,Misdirected
-        Request,\"[RFC7540, Section 9.1.2]\"\r\n422,Unprocessable Entity,[RFC4918]\r\n423,Locked,[RFC4918]\r\n424,Failed
-        Dependency,[RFC4918]\r\n425,Unassigned,\r\n426,Upgrade Required,\"[RFC7231,
-        Section 6.5.15]\"\r\n427,Unassigned,\r\n428,Precondition Required,[RFC6585]\r\n429,Too
-        Many Requests,[RFC6585]\r\n430,Unassigned,\r\n431,Request Header Fields Too
-        Large,[RFC6585]\r\n432-499,Unassigned,\r\n500,Internal Server Error,\"[RFC7231,
-        Section 6.6.1]\"\r\n501,Not Implemented,\"[RFC7231, Section 6.6.2]\"\r\n502,Bad
-        Gateway,\"[RFC7231, Section 6.6.3]\"\r\n503,Service Unavailable,\"[RFC7231,
-        Section 6.6.4]\"\r\n504,Gateway Timeout,\"[RFC7231, Section 6.6.5]\"\r\n505,HTTP
-        Version Not Supported,\"[RFC7231, Section 6.6.6]\"\r\n506,Variant Also Negotiates,[RFC2295]\r\n507,Insufficient
-        Storage,[RFC4918]\r\n508,Loop Detected,[RFC5842]\r\n509,Unassigned,\r\n510,Not
-        Extended,[RFC2774]\r\n511,Network Authentication Required,[RFC6585]\r\n512-599,Unassigned,\r\n"
-    http_version: 
-  recorded_at: Thu, 05 Nov 2015 15:23:25 GMT
-recorded_with: VCR 3.0.0
